Renaming sweep_namespace.h to sweep.h. Simplifying sweep.h. Deprecating spls/spls.h and moving contents to sweep.h
