+++ modules/linear_boltzmann_solvers/b_discrete_ordinates_solver/sweep/angle_set/angle_set.h
@@ -11,6 +11,7 @@ namespace opensn
 namespace lbs
 {
 
+class SweepChunk;
 typedef SweepBoundary SweepBndry;
 
 class AngleSet
