+++ modules/linear_boltzmann_solvers/b_discrete_ordinates_solver/sweep/spds/spds.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include "modules/linear_boltzmann_solvers/b_discrete_ordinates_solver/sweep/spls/spls.h"
-#include "modules/linear_boltzmann_solvers/b_discrete_ordinates_solver/sweep/sweep_namespace.h"
+#include "modules/linear_boltzmann_solvers/b_discrete_ordinates_solver/sweep/sweep.h"
 #include <memory>
 
 namespace opensn
